commit me now pls
If we use "replace_path" in the .mod file but the folder is not in the exact same case as the original the game will load both datasets, vanilla and mod. If the case matches then only the mod files will be loaded. If we don't define the folder in the mod at all then the vanilla wont be loaded, so now all BICE province history folders have a prefix